feat(3ds): Implement New 3DS citro3d renderer with 2D UI rendering, touch input, and memory management
Adds a working PICA200 rendering backend (Citro3dBackend) for the New
Nintendo 3DS port, covering the full 2D UI/menu/HUD draw path via
Render2DClass: shader/texture/vertex-buffer handling, tiled texture
upload with per-format PICA200 swizzling (R5G6B5/A4R4G4B4/A1R5G5B5 at
native 2-byte-per-pixel formats, no RGBA8 bloat), and font glyph
rendering using the existing embedded-font pipeline shared with Switch.

Also includes:
- Touch input wired to the game's mouse path (finger events -> left
  click, absolute-position capture mode) matching the existing
  Android/Switch pattern.
- New3DS heap sizing via the process resource limit API (the same
  mechanism libctru's own default allocator uses) instead of guessing
  fixed sizes.
- Root-caused and fixed a match-load out-of-memory crash: StubD3D8Device
  was permanently double-storing every texture's pixel data (CPU-side
  scratch buffer plus the GPU-side upload); CPU-side scratch is now
  released after a texture's one-time GPU upload for write-once
  textures (everything except the incrementally-rebuilt font glyph
  atlas, which needs its scratch to persist across re-locks).
  GameClient::preloadAssets' whole-faction-roster asset warm-up pass is
  skipped on 3DS for the same reason.
- Fixed a null-pointer crash in W3DTreeBuffer::updateVertexBuffer for
  deleted/missing tree types, matching the existing guard already used
  by its sibling function.
- Various boot/build fixes: devkitARM toolchain file, PICA vertex
  shader compilation via picasso + a generic bin2h embedding script,
  stack size and heap allocation overrides, and window/display
  placement fixes for the bottom (touch-capable) screen.

+++ Generals/Code/GameEngine/Source/Common/System/SaveGame/GameState.cpp
@@ -1226,7 +1226,7 @@ void GameState::populateSaveGameListbox( GameWindow *listbox, SaveLoadLayoutType
 	}
 
 	// select the top "new game" entry
-	GadgetListBoxSetSelected( listbox, 0 );
+	GadgetListBoxSetSelected( listbox, (Int)0 );
 
 }
