+++ GeneralsMD/Code/GameEngine/Source/GameClient/GUI/GUICallbacks/Menus/ReplayMenu.cpp
@@ -383,7 +383,7 @@ void PopulateReplayFileListbox(GameWindow *listbox)
 				break;
 		}
 	}
-	GadgetListBoxSetSelected(listbox, 0);
+	GadgetListBoxSetSelected(listbox, (Int)0);
 }
 
 //-------------------------------------------------------------------------------------------------
